+++ src/fr/zunf1x/mc2d/game/level/blocks/BlockGrass.java
@@ -1,5 +1,6 @@
 package fr.zunf1x.mc2d.game.level.blocks;
 
+import fr.zunf1x.mc2d.math.vectors.Vector2d;
 import fr.zunf1x.mc2d.math.vectors.Vector2f;
 import fr.zunf1x.mc2d.rendering.Color4f;
 import fr.zunf1x.mc2d.rendering.Renderer;
@@ -14,7 +15,7 @@ public class BlockGrass extends Block {
     }
 
     @Override
-    public void render(Vector2f loc, Color4f grassColor) {
+    public void render(Vector2d loc, Color4f grassColor) {
         Texture.BLOCKS.bind();
         Renderer.directTexturedCube(loc.getX(), loc.getY(), new Color4f(1, 1, 1), 2);
         Renderer.directTexturedCube(loc.getX(), loc.getY(), grassColor, 3);
