+++ src/fr/zunf1x/mc2d/game/level/blocks/BlockLeaves.java
@@ -1,5 +1,7 @@
 package fr.zunf1x.mc2d.game.level.blocks;
 
+import fr.zunf1x.mc2d.math.Mathf;
+import fr.zunf1x.mc2d.math.vectors.Vector2d;
 import fr.zunf1x.mc2d.math.vectors.Vector2f;
 import fr.zunf1x.mc2d.rendering.Color4f;
 import fr.zunf1x.mc2d.rendering.Renderer;
@@ -12,7 +14,7 @@ public class BlockLeaves extends Block {
     }
 
     @Override
-    public void render(Vector2f loc, Color4f grassColor) {
+    public void render(Vector2d loc, Color4f grassColor) {
         Texture.BLOCKS.bind();
         Renderer.directTexturedCube(loc.getX(), loc.getY(), grassColor, 14);
         Texture.BLOCKS.unbind();
